agents/pyproject.toml
[project]
name = "agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.0.13",
    "fastapi>=0.104.1",
    "fastmcp>=2.8.0",
    "langchain>=0.3.25",
    "langchain-community>=0.3.25",
    "langchain-google-genai>=2.1.5",
    "langchain-mcp-adapters>=0.1.7",
    "langchain-openai>=0.3.22",
    "langgraph>=0.4.8",
    "langgraph-supervisor>=0.0.27",
    "langsmith>=0.3.45",
    "mcp[cli]>=1.2.1",
    "opik>=1.7.39",
    "pypdf>=5.6.1",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "python-dotenv>=0.19.0",
    "ragas>=0.2.15",
    "typing-extensions>=4.13.1",
    "uvicorn>=0.24.0",
]


